Remove Thrift headers from Jaeger public headers
Adresses https://github.com/jaegertracing/jaeger-client-cpp/issues/121
so that thrift is a private dependency of jaeger.
