+++ src/jaegertracing/Reference.cpp
@@ -15,3 +15,27 @@
  */
 
 #include "jaegertracing/Reference.h"
+#include "jaegertracing/thrift-gen/jaeger_types.h"
+
+namespace jaegertracing {
+void Reference::thrift(thrift::SpanRef& spanRef) const
+{
+    switch (_type) {
+    case Type::ChildOfRef: {
+        spanRef.__set_refType(thrift::SpanRefType::CHILD_OF);
+    } break;
+    case Type::FollowsFromRef: {
+        spanRef.__set_refType(thrift::SpanRefType::FOLLOWS_FROM);
+    } break;
+    default: {
+        std::ostringstream oss;
+        oss << "Invalid span reference type " << static_cast<int>(_type)
+            << ", context " << _spanContext;
+        throw std::invalid_argument(oss.str());
+    } break;
+    }
+    spanRef.__set_traceIdHigh(_spanContext.traceID().high());
+    spanRef.__set_traceIdLow(_spanContext.traceID().low());
+    spanRef.__set_spanId(_spanContext.spanID());
+}
+}  // namespace jaegertracing
