+++ src/jaegertracing/Reference.h
@@ -24,9 +24,11 @@
 
 #include "jaegertracing/Compilers.h"
 #include "jaegertracing/SpanContext.h"
-#include "jaegertracing/thrift-gen/jaeger_types.h"
 
 namespace jaegertracing {
+namespace thrift {
+class SpanRef;
+}
 
 class Reference {
   public:
@@ -42,28 +44,7 @@ class Reference {
 
     Type type() const { return _type; }
 
-    thrift::SpanRef thrift() const
-    {
-        thrift::SpanRef spanRef;
-        switch (_type) {
-        case Type::ChildOfRef: {
-            spanRef.__set_refType(thrift::SpanRefType::CHILD_OF);
-        } break;
-        case Type::FollowsFromRef: {
-            spanRef.__set_refType(thrift::SpanRefType::FOLLOWS_FROM);
-        } break;
-        default: {
-            std::ostringstream oss;
-            oss << "Invalid span reference type " << static_cast<int>(_type)
-                << ", context " << _spanContext;
-            throw std::invalid_argument(oss.str());
-        } break;
-        }
-        spanRef.__set_traceIdHigh(_spanContext.traceID().high());
-        spanRef.__set_traceIdLow(_spanContext.traceID().low());
-        spanRef.__set_spanId(_spanContext.spanID());
-        return spanRef;
-    }
+    void thrift(thrift::SpanRef& spanRef) const;
 
   private:
     SpanContext _spanContext;
