+++ src/jaegertracing/Tag.h
@@ -19,7 +19,6 @@
 
 #include "jaegertracing/Compilers.h"
 
-#include "jaegertracing/thrift-gen/jaeger_types.h"
 #include <algorithm>
 #include <cstdint>
 #include <opentracing/string_view.h>
@@ -28,6 +27,9 @@
 #include <string>
 
 namespace jaegertracing {
+namespace thrift {
+class Tag;
+}
 
 class Tag {
   public:
@@ -41,7 +43,7 @@ class Tag {
     }
 
     template <typename ValueArg>
-    Tag(const std::pair<std::string,ValueArg> & tag_pair)
+    Tag(const std::pair<std::string, ValueArg>& tag_pair)
         : _key(tag_pair.first)
         , _value(tag_pair.second)
     {
@@ -56,67 +58,9 @@ class Tag {
 
     const ValueType& value() const { return _value; }
 
-    thrift::Tag thrift() const
-    {
-        thrift::Tag tag;
-        tag.__set_key(_key);
-        ThriftVisitor visitor(tag);
-        opentracing::util::apply_visitor(visitor, _value);
-        return tag;
-    }
+    void thrift(thrift::Tag& tag) const;
 
   private:
-    class ThriftVisitor {
-      public:
-        using result_type = void;
-
-        explicit ThriftVisitor(thrift::Tag& tag)
-            : _tag(tag)
-        {
-        }
-
-        void operator()(const std::string& value) const { setString(value); }
-
-        void operator()(const char* value) const { setString(value); }
-
-        void operator()(double value) const
-        {
-            _tag.__set_vType(thrift::TagType::DOUBLE);
-            _tag.__set_vDouble(value);
-        }
-
-        void operator()(bool value) const
-        {
-            _tag.__set_vType(thrift::TagType::BOOL);
-            _tag.__set_vBool(value);
-        }
-
-        void operator()(int64_t value) const { setLong(value); }
-
-        void operator()(uint64_t value) const { setLong(value); }
-
-        template <typename Arg>
-        void operator()(Arg&& value) const
-        {
-            // No-op
-        }
-
-      private:
-        void setString(opentracing::string_view value) const
-        {
-            _tag.__set_vType(thrift::TagType::STRING);
-            _tag.__set_vStr(value);
-        }
-
-        void setLong(int64_t value) const
-        {
-            _tag.__set_vType(thrift::TagType::LONG);
-            _tag.__set_vLong(value);
-        }
-
-        thrift::Tag& _tag;
-    };
-
     std::string _key;
     ValueType _value;
 };
