+++ src/jaegertracing/reporters/Config.cpp
@@ -15,12 +15,33 @@
  */
 
 #include "jaegertracing/reporters/Config.h"
+#include "jaegertracing/UDPTransport.h"
+#include "jaegertracing/reporters/CompositeReporter.h"
+#include "jaegertracing/reporters/LoggingReporter.h"
+#include "jaegertracing/reporters/RemoteReporter.h"
 
 namespace jaegertracing {
 namespace reporters {
 
 constexpr int Config::kDefaultQueueSize;
 constexpr const char* Config::kDefaultLocalAgentHostPort;
+
+std::unique_ptr<Reporter> Config::makeReporter(const std::string& serviceName,
+                                               logging::Logger& logger,
+                                               metrics::Metrics& metrics) const
+{
+    std::unique_ptr<UDPTransport> sender(
+        new UDPTransport(net::IPAddress::v4(_localAgentHostPort), 0));
+    std::unique_ptr<RemoteReporter> remoteReporter(new RemoteReporter(
+        _bufferFlushInterval, _queueSize, std::move(sender), logger, metrics));
+    if (_logSpans) {
+        logger.info("Initializing logging reporter");
+        return std::unique_ptr<CompositeReporter>(new CompositeReporter(
+            { std::shared_ptr<RemoteReporter>(std::move(remoteReporter)),
+              std::make_shared<LoggingReporter>(logger) }));
+    }
+    return std::unique_ptr<Reporter>(std::move(remoteReporter));
+}
 
 }  // namespace reporters
 }  // namespace jaegertracing
