+++ src/jaegertracing/reporters/Config.h
@@ -22,11 +22,7 @@
 #include <string>
 
 #include "jaegertracing/Logging.h"
-#include "jaegertracing/UDPTransport.h"
 #include "jaegertracing/metrics/Metrics.h"
-#include "jaegertracing/reporters/CompositeReporter.h"
-#include "jaegertracing/reporters/LoggingReporter.h"
-#include "jaegertracing/reporters/RemoteReporter.h"
 #include "jaegertracing/reporters/Reporter.h"
 #include "jaegertracing/utils/YAML.h"
 
@@ -87,24 +83,7 @@ class Config {
 
     std::unique_ptr<Reporter> makeReporter(const std::string& serviceName,
                                            logging::Logger& logger,
-                                           metrics::Metrics& metrics) const
-    {
-        std::unique_ptr<UDPTransport> sender(
-            new UDPTransport(net::IPAddress::v4(_localAgentHostPort), 0));
-        std::unique_ptr<RemoteReporter> remoteReporter(
-            new RemoteReporter(_bufferFlushInterval,
-                               _queueSize,
-                               std::move(sender),
-                               logger,
-                               metrics));
-        if (_logSpans) {
-            logger.info("Initializing logging reporter");
-            return std::unique_ptr<CompositeReporter>(new CompositeReporter(
-                { std::shared_ptr<RemoteReporter>(std::move(remoteReporter)),
-                  std::make_shared<LoggingReporter>(logger) }));
-        }
-        return std::unique_ptr<Reporter>(std::move(remoteReporter));
-    }
+                                           metrics::Metrics& metrics) const;
 
     int queueSize() const { return _queueSize; }
